As a first step to allow custom ranges allow RangeTrainer to accept a range as props
