+++ App.js
@@ -1,9 +1,10 @@
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
 import RangeTrainer from "./RangeTrainer/RangeTrainer";
+import positionWithGroups from "./RangeTrainer/positionWithGroups";
 
 export default class App extends React.Component {
   render() {
-    return <RangeTrainer />;
+    return <RangeTrainer positionWithGroups={positionWithGroups} />;
   }
 }
